# docs/swagger.yaml
basePath: /
definitions:
  hotel-management_internal_dto.BookingHistoryResponse:
    properties:
      end_date:
        type: string
      id:
        type: integer
      is_paid:
        type: boolean
      rooms:
        items:
          $ref: '#/definitions/hotel-management_internal_dto.BookingHistoryRoom'
        type: array
      start_date:
        type: string
      status:
        type: string
      total_price:
        type: number
    type: object
  hotel-management_internal_dto.BookingHistoryRoom:
    properties:
      bed_num:
        type: integer
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      type:
        type: string
    type: object
  hotel-management_internal_dto.CreateBookingRequest:
    properties:
      end_date:
        type: string
      room_ids:
        items:
          type: integer
        minItems: 1
        type: array
      start_date:
        type: string
    required:
    - end_date
    - room_ids
    - start_date
    type: object
...
  hotel-management_internal_dto.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  hotel-management_internal_dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
...
  hotel-management_internal_dto.RefreshTokenInput:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  hotel-management_internal_dto.RegisterRequest:
    properties:
      email:
        type: string
      first_name:
        minLength: 1
        type: string
      last_name:
        minLength: 1
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    type: object
  hotel-management_internal_dto.Response:
    properties:
      data: {}
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is an API for hotel management.
  title: Hotel Management API
  version: "1.0"
paths:
  /auth/google/callback:
    get:
      consumes:
      - application/json
      description: Handle Google OAuth2 callback, exchange code for tokens, fetch
        user info, create user if not exists, and return JWT tokens.
      parameters:
      - description: Authorization code from Google
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Login successful!
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Code not found from Google.
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to exchange token from Google. / Failed to get user
            information from Google. / Failed to get user / Failed to create user
            / Could not generate token.
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Google OAuth2 callback
      tags:
      - Auth
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return access and refresh tokens
      parameters:
      - description: User login credentials
        in: body
        name: loginRequest
        required: true
        schema:
          $ref: '#/definitions/hotel-management_internal_dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful!
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request data
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Invalid email or password
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Could not generate token
          schema:
            additionalProperties: true
            type: object
      summary: Login
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Get a new access token using a valid refresh token
      parameters:
      - description: Refresh token input
        in: body
        name: refreshTokenInput
        required: true
        schema:
          $ref: '#/definitions/hotel-management_internal_dto.RefreshTokenInput'
      produces:
      - application/json
      responses:
        "200":
          description: New access token generated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request data
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Invalid or expired refresh token
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Could not generate token
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh access token
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with email and password
      parameters:
      - description: User registration data
        in: body
        name: registerRequest
        required: true
        schema:
          $ref: '#/definitions/hotel-management_internal_dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Register user successful!
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Email already exists or invalid request data
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to hash password or create user
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - Auth
  /bookings:
    post:
      consumes:
      - application/json
      description: Allow a customer to create a booking with selected rooms and dates
      parameters:
      - description: Booking request payload
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/hotel-management_internal_dto.CreateBookingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Response with booking details and payment URL
          schema:
            $ref: '#/definitions/hotel-management_internal_dto.Response'
        "400":
          description: Invalid request data/Start date must be before end date/Room
            is not available/Invalid client IP address
          schema:
            $ref: '#/definitions/hotel-management_internal_dto.ErrorResponse'
        "401":
          description: Unauthorized access
          schema:
            $ref: '#/definitions/hotel-management_internal_dto.ErrorResponse'
        "404":
          description: Room not found
          schema:
            $ref: '#/definitions/hotel-management_internal_dto.ErrorResponse'
        "500":
          description: Failed to get room price/Failed to create booking/Failed to
            commit transaction
          schema:
            $ref: '#/definitions/hotel-management_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new booking
      tags:
      - Booking
  /bookings/{id}/cancel:
    get:
      description: Cancel a booking by ID if allowed
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Booking cancelled
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request data
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Booking not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to cancel booking
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Cancel a booking
      tags:
      - Booking
  /bookings/history:
    get:
      description: Retrieve a list of past bookings for the authenticated customer
      produces:
      - application/json
      responses:
        "200":
          description: List of booking history
          schema:
            items:
              $ref: '#/definitions/hotel-management_internal_dto.BookingHistoryResponse'
            type: array
        "401":
          description: Unauthorized access
          schema:
            $ref: '#/definitions/hotel-management_internal_dto.ErrorResponse'
        "500":
          description: Failed to get booking history
          schema:
            $ref: '#/definitions/hotel-management_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get booking history for current customer
      tags:
      - Booking
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"